application/pyproject.toml
[tool.poetry]
name = "fastapi-tutorial"
version = "0.1.0"
description = ""
authors = ["shun198"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11.8"
fastapi = "^0.115.0"
psycopg2 = "^2.9.9"
uvicorn= "^0.17.6"
SQLAlchemy = "^2.0.28"
gunicorn = "^20.1.0"
requests = "^2.31.0"
boto3 = "^1.26.22"
pydantic= "^2.10.6"
sqlalchemy-utils = "^0.41.1"
alembic = "^1.13.1"
email-validator = "^2.1.1"
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.28.1"
pydantic-settings = "^2.9.1"

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.2.1"
freezegun = "^1.2.2"
pytest = "^8.1.1"
pytest-clarity = "^1.0.1"
pytest-cov = "^4.0.0"
pytest-custom-exit-code = "^0.3.0"
pytest-github-actions-annotate-failures = "^0.2.0"
pytest-mock = "^3.10.0"
pytest-sugar = "^1.0.0"
pytest-xdist = "^3.0.2"
ruff = "^0.9.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = "tests"
